update(fix): fix banner download gone
