+++ src/lib/components/ui/InstallBanner.svelte
@@ -5,26 +5,7 @@
     import Label from "$lib/elements/Label.svelte"
     import { Appearance, Shape } from "$lib/enums"
     import Controls from "$lib/layouts/Controls.svelte"
-
-    enum Platform {
-        Windows = "Windows",
-        MacOS = "MacOS",
-        Android = "Android",
-        iOS = "iOS",
-        Linux = "Linux",
-        Other = "Other",
-    }
-
-    function detectPlatform(): Platform {
-        const userAgent = navigator.userAgent.toLowerCase()
-
-        if (userAgent.includes("windows")) return Platform.Windows
-        if (userAgent.includes("mac")) return Platform.MacOS
-        if (userAgent.includes("android")) return Platform.Android
-        if (/iphone|ipad|ipod/.test(userAgent)) return Platform.iOS
-        if (userAgent.includes("linux")) return Platform.Linux
-        return Platform.Other
-    }
+    import { detectPlatform, Platform } from "$lib/utils/DetectPlatform"
 
     function isElectron(): boolean {
         return typeof window !== "undefined" && window.process?.versions?.electron !== undefined
